+++ src/readonlyKeywordRule.ts
@@ -10,19 +10,36 @@ export class Rule extends Lint.Rules.AbstractRule {
   public apply(sourceFile: ts.SourceFile): Lint.RuleFailure[] {
     return this.applyWithFunction(
       sourceFile,
-      (ctx: Lint.WalkContext<Shared.Options>) => Shared.walk(ctx, checkNode, "A readonly modifier is required."),
-      Shared.parseOptions(this.ruleArguments));
+      (ctx: Lint.WalkContext<Shared.Options>) =>
+        Shared.walk(ctx, checkNode, "A readonly modifier is required."),
+      Shared.parseOptions(this.ruleArguments)
+    );
   }
 }
 
-function checkNode(node: ts.Node, ctx: Lint.WalkContext<Shared.Options>): ReadonlyArray<Shared.InvalidNode> {
+function checkNode(
+  node: ts.Node,
+  ctx: Lint.WalkContext<Shared.Options>
+): ReadonlyArray<Shared.InvalidNode> {
   return checkPropertySignatureAndIndexSignature(node, ctx);
 }
 
-function checkPropertySignatureAndIndexSignature(node: ts.Node, ctx: Lint.WalkContext<Shared.Options>): ReadonlyArray<Shared.InvalidNode> {
-
-  if (node.kind === ts.SyntaxKind.PropertySignature || node.kind === ts.SyntaxKind.IndexSignature) {
-    if (!(node.modifiers && node.modifiers.filter((m) => m.kind === ts.SyntaxKind.ReadonlyKeyword).length > 0)) {
+function checkPropertySignatureAndIndexSignature(
+  node: ts.Node,
+  ctx: Lint.WalkContext<Shared.Options>
+): ReadonlyArray<Shared.InvalidNode> {
+  if (
+    node.kind === ts.SyntaxKind.PropertySignature ||
+    node.kind === ts.SyntaxKind.IndexSignature ||
+    node.kind === ts.SyntaxKind.PropertyDeclaration
+  ) {
+    if (
+      !(
+        node.modifiers &&
+        node.modifiers.filter(m => m.kind === ts.SyntaxKind.ReadonlyKeyword)
+          .length > 0
+      )
+    ) {
       // Check if ignore-prefix applies
       if (Shared.shouldIgnorePrefix(node, ctx.options, ctx.sourceFile)) {
         return [];
@@ -30,9 +47,17 @@ function checkPropertySignatureAndIndexSignature(node: ts.Node, ctx: Lint.WalkCo
       const length = node.getWidth(ctx.sourceFile);
       // const fulltext = node.getText(ctx.sourceFile);
       const fulltext = node.getText(ctx.sourceFile);
-      return [Shared.createInvalidNode(node, new Lint.Replacement(node.end - length, length, `readonly ${fulltext}`))];
+      return [
+        Shared.createInvalidNode(
+          node,
+          new Lint.Replacement(
+            node.end - length,
+            length,
+            `readonly ${fulltext}`
+          )
+        )
+      ];
     }
   }
   return [];
-
 }